added testing of 'write_to_cert' method

+++ av_certs.rb
@@ -71,7 +71,7 @@ def send_mail(name, email, file)
 end
 
 def save_certificate(name, date, course_name, course_desc)
-  Certificate.where(student_name: name).destroy_all
+  Certificate.where(student_name: name).destroy_all unless Sinatra::Base.production?
   @cert = Certificate.create(student_name: name,
                             generated_at: date,
                             course_name: course_name,
